Stop spawning multiple times per key press
Don't spawn a new tile if the action is "None". That's just silly. Only do it
if the action is a slide. Still need to refine this more so that we only spawn
if a slide actually happens (if no sliding occurs it's not very nice to add a
new tile anyway).