Fix tile sometimes not spawning and overeager collapse
Turns out that the range for the random number generator is inclusive, i.e.,
[start, end], so the range to pick from a list should be [0, len-1], not
[0, len]. This bug caused tiles to occasionally not spawn.

In the fold left to collapse tiles, I was checking the square to the right and
always collapsing if the values matched. This could cause something like
[4, 2, 2] to collapse to [8] in one step instead of to [4, 4]. To resolve,
the fold will combine items into pairs, so [4, 2, 2] becomes [(0, 4), (2, 2)].
At the end, each pair is summed. It's not pretty, and it needs more testing,
but it seems to work.